requirements.txt
fastapi[all]
Pillow==11.1.0
numpy==2.0.2
SimpleITK==2.4.1
pandas==2.2.3
openpyxl==3.1.5
pathlib==1.0.1
python-dotenv==1.1.1
pynrrd==1.1.3
nibabel==5.3.2
scikit-image==0.24.0
psycopg2==2.9.10
requests==2.32.4
python-irodsclient==3.0.0
nninteractive==1.1.2
huggingface_hub==0.34.3